move upgrade_manifest_json + Documentation to dbt/artifacts
