+++ core/dbt/artifacts/manifest/__init__.py
@@ -1,0 +1,1 @@
+from dbt.artifacts.manifest.manifest import WritableManifest, ManifestMetadata, UniqueID
