+++ core/dbt/artifacts/manifest/upgrade.py
