+++ core/dbt/artifacts/nodes/__init__.py
@@ -1,0 +1,2 @@
+from dbt.artifacts.nodes.base import BaseArtifactNode
+from dbt.artifacts.nodes.documentation import Documentation
